% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ad_spending_overall}
\alias{ad_spending_overall}
\title{Presidental Election Adverstisment Spending}
\format{
A data frame with two variables: \code{Year},
\code{Amount Spent on Ads}
}
\source{
\url{https://www.kaggle.com/fec/independent-political-ad-spending}
}
\usage{
data(ad_spending_overall)
}
\description{
Amount spent on advertisments for the 2004, 08, 12, and 16
presidential elections
}
\keyword{datasets}
