% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ad_spending_total}
\alias{ad_spending_total}
\title{Presidental Election Ad Spending by Party}
\format{
A data frame with three variables: \code{Year},
\code{Amount Spent on Ads}, \code{Party}
}
\source{
\url{https://www.kaggle.com/fec/independent-political-ad-spending}
}
\usage{
data(ad_spending_total)
}
\description{
Amount spent on advertisments by politcal parties for the
2004, 08, 12, and 16 presidential elections
}
\keyword{datasets}
