% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ad_support_oppose}
\alias{ad_support_oppose}
\title{Detailed Presidental Election Adverstisment Spending}
\format{
A data frame with four variables: \code{Year},
\code{Amount Spent on Ads}, \code{Party}, \code{Position}
}
\source{
\url{https://www.kaggle.com/fec/independent-political-ad-spending}
}
\usage{
data(ad_support_oppose)
}
\description{
Amount spent on advertisments for the 2004, 08, 12, and 16
presidential elections in support of and opposing each of the major
party candiates
}
\keyword{datasets}
