% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_counties}
\alias{all_counties}
\title{USA county names}
\format{
A data frame with three variables: \code{county}, \code{state_name},
\code{state_abb}{state abbriviation}
}
\source{
\url{https://eric.clst.org/tech/usgeojson/}

\url{https://www.census.gov/}
}
\usage{
data(all_counties)
}
\description{
USA county names
}
\keyword{datasets}
