% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cal_poly_stalking}
\alias{cal_poly_stalking}
\title{Cal Poly Stalking Data}
\format{
A data frame with four variables: \code{Year},
\code{Proportion of Students}{prop of students who reported experiencing this type of assult},
\code{University},  \code{Offense}
}
\source{
University Cleary Reports
}
\usage{
data(cal_poly_stalking)
}
\description{
Information on stalking occurances from 2014-18 at Cal Poly
}
\keyword{datasets}
