% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_choropleths.R
\name{check_county_names}
\alias{check_county_names}
\title{Check your data for use in 'county_map()` function}
\usage{
check_county_names(data, state = "all")
}
\arguments{
\item{data}{Your data frame to make a choropleth. Contains three columns,
county name, county state and your variable of intrest, in that order}

\item{state}{A string contianing the state name or abbriviation that
you would like to map. Default is "all" which maps the entire USA}
}
\value{
A data frame or a list identifying which counties are in the function
data set, but were not found in yours, and which counties are in your data set
but were not found in the function's. Please change the names of the counties
in your data set to match those in the function if you'd like to graph those
counties
}
\description{
Checks your dataframe of counties against the county names in the
\code{county_map()} to identify any missspellings or spacing issues
}
\examples{
check_county_names(median_income)

}
