% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greek_life.R
\name{comp_gl_sanctions}
\alias{comp_gl_sanctions}
\title{Compare CP Greek Life sanction information}
\usage{
comp_gl_sanctions(old, new = "current")
}
\arguments{
\item{old}{A chapter sanction information data frame with 7 columns created
by \code{get_gl_sanctions()}}

\item{new}{A chapter sanction information data frame with 7 columns created
by \code{get_gl_sanctions()}. Default is "current" which scrapes the Greek Life
Sanctions website for the current sanction list}
}
\value{
An html table. Green indicates information that has been added and
red indicates information that has been removed from the site.
}
\description{
Compares old Cal Poly Greek Life sanction information with
what is currently posted on the website
}
\examples{
comp_gl_sanctions(old_sanctions)

}
\references{
\url{https://greeklife.calpoly.edu/chapter-sanction-information}
}
