% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_choropleths.R
\name{county_map}
\alias{county_map}
\title{Creates a choropleth of USA by county}
\usage{
county_map(data, state = "all")
}
\arguments{
\item{data}{Your data frame to make a choropleth. Contains three columns,
county name, county state and your variable of intrest, in that order}

\item{state}{A string contianing the state name or abbriviation that
you would like to map. Default is "all" which maps the entire USA}
}
\value{
An interactive map, which is a HTML widget object
}
\description{
Creates a choropleth of USA by county
}
\examples{
county_map(median_income, state = "CA")

}
