% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greek_life.R
\name{get_gl_sanctions}
\alias{get_gl_sanctions}
\title{Get CP Greek Life sanction information}
\usage{
get_gl_sanctions(save = FALSE)
}
\arguments{
\item{save}{Saves the outputted data frame as a csv file named
"cpgl_sanctions_MMDDYY.csv" to your current directory. Default is FALSE}
}
\value{
A data frame with 7 columns containing chapter sanction information
}
\description{
Scrapes Cal Poly Greek Life's chapter sanction webpage to get
data frame of current sanctions
}
\examples{

get_gl_sanctions()
## To save the outputted data frame of sanction info as a csv
\dontrun{
get_gl_sanctions(save = TRUE)
}

}
\references{
\url{https://greeklife.calpoly.edu/chapter-sanction-information}
}
