% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_exp}
\alias{life_exp}
\title{USA Life Expectancy}
\format{
A data frame with two variables: \code{state},
\code{Life Expectancy}
}
\source{
state.x77
}
\usage{
data(life_exp)
}
\description{
USA Life Expectancy
}
\keyword{datasets}
