% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{median_income}
\alias{median_income}
\title{USA Median Household Income for 2018}
\format{
A data frame with three variables: \code{Name}{County name},
\code{Postal Code}{State name}, \code{Median Household Income}
}
\source{
\url{https://www.census.gov/data/datasets/2018/demo/saipe/2018-state-and-county.html}
}
\usage{
data(median_income)
}
\description{
USA Median Household Income for 2018
}
\keyword{datasets}
