% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{old_sanctions}
\alias{old_sanctions}
\title{Cal Poly Greek Life Sanctions}
\format{
A data frame with seven variables: \code{Letters},
\code{Chapter}, \code{Additional_Info_1},  \code{Additional_Info_2},
\code{Additional_Info_3},  \code{Additional_Info_4}, \code{Section}
}
\source{
\url{https://web.archive.org/web/20200425094029/https://greeklife.calpoly.edu/chapter-sanction-information}
}
\usage{
data(old_sanctions)
}
\description{
Greek life Sanctions from pulled from website archive 4/25/2020
}
\keyword{datasets}
