% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_plots.R
\name{see_colors}
\alias{see_colors}
\title{See colors in the Cal Poly and Mustang News color palettes}
\source{
\url{https://universitymarketing.calpoly.edu/brand-guidelines/colors/}
}
\usage{
see_colors(palette)
}
\arguments{
\item{palette}{A string indicating which color palette to see.
Options are "cp" for the Cal Poly palette or "mn" for the
Mustang News palette}
}
\value{
An image of all the colors in the palette
}
\description{
See colors in the Cal Poly and Mustang News color palettes
}
