% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_plots.R
\name{set_colors}
\alias{set_colors}
\title{Choose colors for graphs}
\usage{
set_colors(palette, graph, ...)
}
\arguments{
\item{palette}{A string indicating which color palette to see.
Options are "cp" for the Cal Poly palette or "mn" for the
Mustang News palette}

\item{graph}{A string indicating which type of graph you made.
Options are "line" or "bar"}

\item{...}{The colors that you want to use, unquoted, seperated
by commas. There should be one color for each group in variable you
want to color by}
}
\description{
Add this function on to an existing line or bar graph
to change the colors of the graph
}
