% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sexual_assault_4}
\alias{sexual_assault_4}
\title{Sexual Assult Data}
\format{
A data frame with four variables: \code{Year},
\code{Proportion of Students}{prop of students who reported experiencing this type of assult},
\code{University},  \code{Offense}
}
\source{
University Cleary Reports
}
\usage{
data(sexual_assault_4)
}
\description{
Sexual assult data from 2014-18 from 4 CA universities
}
\keyword{datasets}
