% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_plots.R
\name{simple_line}
\alias{simple_line}
\title{Create a simple line graph}
\usage{
simple_line(data, title, subtitle = NULL, caption = NULL, grid = FALSE)
}
\arguments{
\item{data}{A data frame with 2-4 columns. Data should be formatted in
the following way: column 1 is the x-axis variable, column 2 is the
y-axis variable, column 3 is the coloring variable, column 4 is the
variable split the data set on to create side by side graphs}

\item{title}{A string containing the title of the graph}

\item{subtitle}{A string containing the subtitle of the graph. Default
is no subtitle}

\item{caption}{string containing the caption of the graph. Default
is no caption}

\item{grid}{Shows major grid lines, default is FALSE}
}
\value{
A ggplot graph
}
\description{
Create a simple line graph
}
