% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slo_pd.R
\name{slopd_log}
\alias{slopd_log}
\title{Turns SLO PD Public Log into a useable dataframe from its original text file
format}
\usage{
slopd_log(current = TRUE, file = "txt", location = FALSE, google_api = "a")
}
\arguments{
\item{current}{Scrapes current log posted on slocity.org. Defualt is TRUE}

\item{file}{A text file of a SLO PD Public Log to convert}

\item{location}{Gets latitiude and longitude for each adress mentioned in
the log. Must have a valid Google API key to use. Default is FALSE}

\item{google_api}{A string of a valid Google API}
}
\value{
A data frame with each row containing the information for one entry
from the log
}
\description{
Turns SLO PD Public Log into a useable dataframe from its original text file
format
}
\examples{
slopd_log()

}
