% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stalking_3}
\alias{stalking_3}
\title{Stalking Data}
\format{
A data frame with four variables: \code{Year},
\code{Proportion of Students}{prop of students who reported experiencing stalking}
}
\source{
University Cleary Reports
}
\usage{
data(stalking_3)
}
\description{
Information on stalking occurances from 2014-18 from 4 CA universities
}
\keyword{datasets}
