% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_choropleths.R
\name{state_map}
\alias{state_map}
\title{Creates a choropleth of USA by county}
\usage{
state_map(data)
}
\arguments{
\item{data}{Your data frame to make a choropleth. Contains three columns,
county name, county state and your variable of intrest, in that order}
}
\value{
An interactive map, which is a HTML widget object
}
\description{
Creates a choropleth of USA by county
}
\examples{
state_map(life_exp)

}
