% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_upd.R
\name{upd_log}
\alias{upd_log}
\title{Scrapes Cal Poly UPD's Activity Log from the department's website and converts
it into a useable dataframe}
\usage{
upd_log(start = "oldest", end = "newest")
}
\arguments{
\item{start}{A string with the date to start collecting reports from. Please
use "YYYY-MM-DD" format. Keep in mind that the online log only goes back
60 days from the current date. Default is "oldest" and starts collecting
from the oldest available report}

\item{end}{A string with the date to finish collecting reports from.Please
use "YYYY-MM-DD" format. Keep in mind that the online log only goes back
60 days from the current date. Default is "newest" and stops collecting
from the most recent report}
}
\value{
A data frame with each row containing the information for one entry
from the log. If a single entry had mulitple inicident types, then there is one
row for each incident type, so take care when reporting total number of
incidents.
}
\description{
Scrapes Cal Poly UPD's Activity Log from the department's website and converts
it into a useable dataframe
}
\examples{
upd_log(start = "2020-05-22")

}
